requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
nest-asyncio
pickleshare
PyJWT
python-decouple
pyOpenSSL
passlib
alembic
pandas
pandas-profiling
ydata-profiling
openpyxl
python-multipart
aiofiles
websocket
websockets
aiofiles
streamlit
streamlit_option_menu
extra-streamlit-components
streamlit-card
bcrypt